langchain_tutorial/pyproject.toml
[tool.poetry]
name = "langchain-tutorial"
version = "0.1.0"
description = ""
authors = ["jojopvcp <jonathan.ndamba@ext.groupepvcp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.8"
langchain-core = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
